Separate stripe into worker and default entrypoints, needs tests for worker
