+++ lib/utils.js
@@ -2,19 +2,6 @@
 const EventEmitter = require('events').EventEmitter;
 const qs = require('qs');
 const crypto = require('crypto');
-// Certain sandboxed environments (our known example right now are CloudFlare
-// Workers) may make `child_process` unavailable. Because `exec` isn't critical
-// to the operation of stripe-node, we handle this unavailability gracefully.
-let exec = null;
-try {
-    exec = require('child_process').exec;
-}
-catch (e) {
-    // @ts-ignore
-    if (e.code !== 'MODULE_NOT_FOUND') {
-        throw e;
-    }
-}
 const OPTIONS_KEYS = [
     'apiKey',
     'idempotencyKey',
@@ -269,29 +256,6 @@ const utils = {
         }
     },
     emitWarning,
-    /**
-     * Node's built in `exec` function sometimes throws outright,
-     * and sometimes has a callback with an error,
-     * depending on the type of error.
-     *
-     * This unifies that interface.
-     */
-    safeExec: (cmd, cb) => {
-        // Occurs if we couldn't load the `child_process` module, which might
-        // happen in certain sandboxed environments like a CloudFlare Worker.
-        if (utils._exec === null) {
-            cb(new Error('exec not available'), null);
-            return;
-        }
-        try {
-            utils._exec(cmd, cb);
-        }
-        catch (e) {
-            cb(e, null);
-        }
-    },
-    // For mocking in tests.
-    _exec: exec,
     isObject: (obj) => {
         const type = typeof obj;
         return (type === 'function' || type === 'object') && !!obj;
@@ -349,14 +313,6 @@ const utils = {
             }
         }
         return n;
-    },
-    determineProcessUserAgentProperties: () => {
-        return typeof process === 'undefined'
-            ? {}
-            : {
-                lang_version: process.version,
-                platform: process.platform,
-            };
     },
     /**
      * Joins an array of Uint8Arrays into a single Uint8Array
