+++ package.json
@@ -65,5 +65,9 @@
     "fix": "yarn lint --fix && ./scripts/updateAPIVersion.js",
     "report": "nyc -r text -r lcov report",
     "coveralls": "cat coverage/lcov.info | ./node_modules/coveralls/bin/coveralls.js"
+  },
+  "exports": {
+    "worker": "./lib/stripe.worker.js",
+    "default": "./lib/stripe.js"
   }
 }
