+++ test/stripe.spec.js
@@ -4,6 +4,7 @@
 
 const testUtils = require('../testUtils');
 const utils = require('../lib/utils');
+const nodeUtils = require('../lib/nodeUtils');
 const Stripe = require('../lib/stripe');
 const stripe = require('../lib/stripe')(testUtils.getUserStripeKey(), 'latest');
 const crypto = require('crypto');
@@ -249,7 +250,7 @@ describe('Stripe Module', function() {
       });
 
       it('gets added to the user-agent', () => {
-        utils.safeExec = (cmd, cb) => {
+        Stripe.safeExec = (cmd, cb) => {
           cb(null, 'foøname');
         };
         return expect(
