+++ test/utils.spec.js
@@ -384,71 +384,6 @@ describe('utils', () => {
     });
   });
 
-  describe('safeExec', () => {
-    let origExec;
-    beforeEach(() => {
-      origExec = utils._exec;
-    });
-    afterEach(() => {
-      utils._exec = origExec;
-    });
-
-    it('runs exec', () => {
-      const calls = [];
-      utils._exec = (cmd, cb) => {
-        calls.push([cmd, cb]);
-      };
-
-      function myCb() {}
-      utils.safeExec('hello', myCb);
-      expect(calls).to.deep.equal([['hello', myCb]]);
-    });
-
-    it('passes along normal errors', () => {
-      const myErr = Error('hi');
-      utils._exec = (cmd, cb) => {
-        cb(myErr, null);
-      };
-
-      const calls = [];
-      function myCb(err, x) {
-        calls.push([err, x]);
-      }
-      utils.safeExec('hello', myCb);
-      expect(calls).to.deep.equal([[myErr, null]]);
-    });
-
-    it('passes along thrown errors as normal callback errors', () => {
-      const myErr = Error('hi');
-      utils._exec = (cmd, cb) => {
-        throw myErr;
-      };
-
-      const calls = [];
-      function myCb(err, x) {
-        calls.push([err, x]);
-      }
-      utils.safeExec('hello', myCb);
-      expect(calls).to.deep.equal([[myErr, null]]);
-    });
-
-    it('handles being unable to require `child_process`', () => {
-      utils._exec = null;
-
-      let actualErr = null;
-      let actualRes = null;
-      function myCb(err, res) {
-        actualErr = err;
-        actualRes = res;
-      }
-      utils.safeExec('hello', myCb);
-      expect(actualErr.toString()).to.equal(
-        new Error('exec not available').toString()
-      );
-      expect(actualRes).to.equal(null);
-    });
-  });
-
   describe('flattenAndStringify', () => {
     it('Stringifies primitive types', () => {
       expect(
